requirements.txt
Flask==3.1.1
feedgen==0.9.0
requests==2.32.4
python-dateutil==2.8.2
